# models/auth_data.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// AuthData Модель данных новой сессии
//
// swagger:model AuthData
type AuthData struct {

	// Электронная почта
	// Max Length: 50
	// Format: email
	Email *strfmt.Email `json:"email,omitempty" sql:"email"`

	// Пароль
	// Min Length: 6
	Password *string `json:"password,omitempty" sql:"password"`

	// Номер телефона
	Phone *int64 `json:"phone,omitempty" sql:"phone"`
}

// Validate validates this auth data
func (m *AuthData) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateEmail(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validatePassword(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *AuthData) validateEmail(formats strfmt.Registry) error {

	if swag.IsZero(m.Email) { // not required
		return nil
	}

	if err := validate.MaxLength("email", "body", string(*m.Email), 50); err != nil {
		return err
	}

	if err := validate.FormatOf("email", "body", "email", m.Email.String(), formats); err != nil {
		return err
	}

	return nil
}

func (m *AuthData) validatePassword(formats strfmt.Registry) error {

	if swag.IsZero(m.Password) { // not required
		return nil
	}

	if err := validate.MinLength("password", "body", string(*m.Password), 6); err != nil {
		return err
	}

	return nil
}

// MarshalBinary interface implementation
func (m *AuthData) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *AuthData) UnmarshalBinary(b []byte) error {
	var res AuthData
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}