# models/basic_user.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// BasicUser Пользователь
//
// swagger:model BasicUser
type BasicUser struct {

	// Дата рождения
	// Format: date
	Birthday *strfmt.Date `json:"birthday,omitempty" sql:"birthday"`

	// Адрес электронной почты
	// Max Length: 50
	// Format: email
	Email *strfmt.Email `json:"email,omitempty" sql:"email"`

	// Имя
	// Max Length: 50
	Name *string `json:"name,omitempty" sql:"name"`

	// Номер телефона
	// Maximum: 7.9999999999e+10
	// Minimum: 7.9000000001e+10
	Phone *int64 `json:"phone,omitempty" sql:"phone"`

	// Отчество
	// Max Length: 50
	Pname *string `json:"pname,omitempty" sql:"pname"`

	// Роль пользователя
	Role *string `json:"role,omitempty" sql:"role"`

	// Фамилия
	// Max Length: 50
	Sname *string `json:"sname,omitempty" sql:"sname"`
}

// Validate validates this basic user
func (m *BasicUser) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateBirthday(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateEmail(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateName(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validatePhone(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validatePname(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateSname(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *BasicUser) validateBirthday(formats strfmt.Registry) error {

	if swag.IsZero(m.Birthday) { // not required
		return nil
	}

	if err := validate.FormatOf("birthday", "body", "date", m.Birthday.String(), formats); err != nil {
		return err
	}

	return nil
}

func (m *BasicUser) validateEmail(formats strfmt.Registry) error {

	if swag.IsZero(m.Email) { // not required
		return nil
	}

	if err := validate.MaxLength("email", "body", string(*m.Email), 50); err != nil {
		return err
	}

	if err := validate.FormatOf("email", "body", "email", m.Email.String(), formats); err != nil {
		return err
	}

	return nil
}

func (m *BasicUser) validateName(formats strfmt.Registry) error {

	if swag.IsZero(m.Name) { // not required
		return nil
	}

	if err := validate.MaxLength("name", "body", string(*m.Name), 50); err != nil {
		return err
	}

	return nil
}

func (m *BasicUser) validatePhone(formats strfmt.Registry) error {

	if swag.IsZero(m.Phone) { // not required
		return nil
	}

	if err := validate.MinimumInt("phone", "body", int64(*m.Phone), 7.9000000001e+10, false); err != nil {
		return err
	}

	if err := validate.MaximumInt("phone", "body", int64(*m.Phone), 7.9999999999e+10, false); err != nil {
		return err
	}

	return nil
}

func (m *BasicUser) validatePname(formats strfmt.Registry) error {

	if swag.IsZero(m.Pname) { // not required
		return nil
	}

	if err := validate.MaxLength("pname", "body", string(*m.Pname), 50); err != nil {
		return err
	}

	return nil
}

func (m *BasicUser) validateSname(formats strfmt.Registry) error {

	if swag.IsZero(m.Sname) { // not required
		return nil
	}

	if err := validate.MaxLength("sname", "body", string(*m.Sname), 50); err != nil {
		return err
	}

	return nil
}

// MarshalBinary interface implementation
func (m *BasicUser) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *BasicUser) UnmarshalBinary(b []byte) error {
	var res BasicUser
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}