# models/edit_user.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// EditUser Данные для изменения пользователя
//
// swagger:model EditUser
type EditUser struct {
	BasicUser

	EditUserAllOf1
}

// UnmarshalJSON unmarshals this object from a JSON structure
func (m *EditUser) UnmarshalJSON(raw []byte) error {
	// AO0
	var aO0 BasicUser
	if err := swag.ReadJSON(raw, &aO0); err != nil {
		return err
	}
	m.BasicUser = aO0

	// AO1
	var aO1 EditUserAllOf1
	if err := swag.ReadJSON(raw, &aO1); err != nil {
		return err
	}
	m.EditUserAllOf1 = aO1

	return nil
}

// MarshalJSON marshals this object to a JSON structure
func (m EditUser) MarshalJSON() ([]byte, error) {
	_parts := make([][]byte, 0, 2)

	aO0, err := swag.WriteJSON(m.BasicUser)
	if err != nil {
		return nil, err
	}
	_parts = append(_parts, aO0)

	aO1, err := swag.WriteJSON(m.EditUserAllOf1)
	if err != nil {
		return nil, err
	}
	_parts = append(_parts, aO1)
	return swag.ConcatJSON(_parts...), nil
}

// Validate validates this edit user
func (m *EditUser) Validate(formats strfmt.Registry) error {
	var res []error

	// validation for a type composition with BasicUser
	if err := m.BasicUser.Validate(formats); err != nil {
		res = append(res, err)
	}
	// validation for a type composition with EditUserAllOf1

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

// MarshalBinary interface implementation
func (m *EditUser) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *EditUser) UnmarshalBinary(b []byte) error {
	var res EditUser
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}

// EditUserAllOf1 edit user all of1
//
// swagger:model EditUserAllOf1
type EditUserAllOf1 interface{}