# models/new_user.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"encoding/json"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// NewUser Данные для создания пользователя
//
// swagger:model NewUser
type NewUser struct {
	BasicUser

	// Пароль пользователя
	//
	// Required: true
	Password *string `json:"password" sql:"password"`

	// Способ регистрации
	// * phone - по номеру телефона * email  - по адресу электронной почты
	//
	// Required: true
	// Enum: [phone email]
	RegisterType *string `json:"registerType" sql:"register_type"`
}

// UnmarshalJSON unmarshals this object from a JSON structure
func (m *NewUser) UnmarshalJSON(raw []byte) error {
	// AO0
	var aO0 BasicUser
	if err := swag.ReadJSON(raw, &aO0); err != nil {
		return err
	}
	m.BasicUser = aO0

	// AO1
	var dataAO1 struct {
		Password *string `json:"password"`

		RegisterType *string `json:"registerType"`
	}
	if err := swag.ReadJSON(raw, &dataAO1); err != nil {
		return err
	}

	m.Password = dataAO1.Password

	m.RegisterType = dataAO1.RegisterType

	return nil
}

// MarshalJSON marshals this object to a JSON structure
func (m NewUser) MarshalJSON() ([]byte, error) {
	_parts := make([][]byte, 0, 2)

	aO0, err := swag.WriteJSON(m.BasicUser)
	if err != nil {
		return nil, err
	}
	_parts = append(_parts, aO0)
	var dataAO1 struct {
		Password *string `json:"password"`

		RegisterType *string `json:"registerType"`
	}

	dataAO1.Password = m.Password

	dataAO1.RegisterType = m.RegisterType

	jsonDataAO1, errAO1 := swag.WriteJSON(dataAO1)
	if errAO1 != nil {
		return nil, errAO1
	}
	_parts = append(_parts, jsonDataAO1)
	return swag.ConcatJSON(_parts...), nil
}

// Validate validates this new user
func (m *NewUser) Validate(formats strfmt.Registry) error {
	var res []error

	// validation for a type composition with BasicUser
	if err := m.BasicUser.Validate(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validatePassword(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateRegisterType(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *NewUser) validatePassword(formats strfmt.Registry) error {

	if err := validate.Required("password", "body", m.Password); err != nil {
		return err
	}

	return nil
}

var newUserTypeRegisterTypePropEnum []interface{}

func init() {
	var res []string
	if err := json.Unmarshal([]byte(`["phone","email"]`), &res); err != nil {
		panic(err)
	}
	for _, v := range res {
		newUserTypeRegisterTypePropEnum = append(newUserTypeRegisterTypePropEnum, v)
	}
}

// property enum
func (m *NewUser) validateRegisterTypeEnum(path, location string, value string) error {
	if err := validate.EnumCase(path, location, value, newUserTypeRegisterTypePropEnum, true); err != nil {
		return err
	}
	return nil
}

func (m *NewUser) validateRegisterType(formats strfmt.Registry) error {

	if err := validate.Required("registerType", "body", m.RegisterType); err != nil {
		return err
	}

	// value enum
	if err := m.validateRegisterTypeEnum("registerType", "body", *m.RegisterType); err != nil {
		return err
	}

	return nil
}

// MarshalBinary interface implementation
func (m *NewUser) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *NewUser) UnmarshalBinary(b []byte) error {
	var res NewUser
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}