# models/session.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// Session Модель сессии
//
// swagger:model Session
type Session struct {

	// CRSF-токен
	// Max Length: 32
	// Min Length: 32
	CRSFToken *string `json:"CRSFToken,omitempty" sql:"csrf_token"`

	// Срок действия
	// Format: date-time
	LifeTime *strfmt.DateTime `json:"lifeTime,omitempty" sql:"life_time"`

	// Сессия
	// Max Length: 32
	// Min Length: 32
	Session *string `json:"session,omitempty" sql:"session"`

	// Идентификатор пользователя
	UserID *int64 `json:"userID,omitempty" sql:"user_id"`
}

// Validate validates this session
func (m *Session) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateCRSFToken(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateLifeTime(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateSession(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *Session) validateCRSFToken(formats strfmt.Registry) error {

	if swag.IsZero(m.CRSFToken) { // not required
		return nil
	}

	if err := validate.MinLength("CRSFToken", "body", string(*m.CRSFToken), 32); err != nil {
		return err
	}

	if err := validate.MaxLength("CRSFToken", "body", string(*m.CRSFToken), 32); err != nil {
		return err
	}

	return nil
}

func (m *Session) validateLifeTime(formats strfmt.Registry) error {

	if swag.IsZero(m.LifeTime) { // not required
		return nil
	}

	if err := validate.FormatOf("lifeTime", "body", "date-time", m.LifeTime.String(), formats); err != nil {
		return err
	}

	return nil
}

func (m *Session) validateSession(formats strfmt.Registry) error {

	if swag.IsZero(m.Session) { // not required
		return nil
	}

	if err := validate.MinLength("session", "body", string(*m.Session), 32); err != nil {
		return err
	}

	if err := validate.MaxLength("session", "body", string(*m.Session), 32); err != nil {
		return err
	}

	return nil
}

// MarshalBinary interface implementation
func (m *Session) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *Session) UnmarshalBinary(b []byte) error {
	var res Session
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}